==> 4_make_tables.ipynb <==
# Visualizations

import numpy as np
import matplotlib.pyplot as plt
import pickle

path_base = "results/real_data"

with open(path_base + "/results.pickle", "rb") as f:
    smaller_results = pickle.load(f)


data_names = ["credit", "adult", "gmsc"]

results = {}
for data_name in data_names:
    results[data_name] = {}
    results[data_name] = {
        "LiCE_optimize": smaller_results[data_name]["LiCE_optimize"],
        "LiCE_median": smaller_results[data_name]["LiCE_median"],
        "LiCE_quartile": smaller_results[data_name]["LiCE_quartile"],
        "LiCE_sample": smaller_results[data_name]["LiCE_sample"],
        "MIO + SPN": smaller_results[data_name]["MIO_no_spn"],

        "VAE + SPN": smaller_results[data_name]["VAE"],
        "DiCE + SPN": smaller_results[data_name]["DiCE"],
        "CH-CVAE ": smaller_results[data_name]["CVAE"],
        "FACE_knn": smaller_results[data_name]["FACE_knn"],
        "FACE_eps": smaller_results[data_name]["FACE_eps"],
        "PROPLACE": smaller_results[data_name]["PROPLACE"],
    }

res_type = "valid"
# res_type = "actionable"
method = "LiCE_optimize"

data_names = ["gmsc", "adult", "credit"]

print("LL value difference")
print("\t" + "\t\t".join(data_names))
print("true ", end="")
for data_name in data_names:
    curr_res = results[data_name][method]
    mean_ll = np.mean([res[res_type]["ll"] for res in curr_res.values() if res_type in res])
    print(f"& \\scinum{{{mean_ll}}} $\\pm$ \\scinum{{{np.std([res[res_type]['ll'] for res in curr_res.values() if res_type in res])}}} ", end="")
print()
print("mio ", end="")
for data_name in data_names:
    curr_res = results[data_name][method]
    mean_ll_comp = np.mean([res[res_type]["comp_ll"] for res in curr_res.values() if res_type in res])
    print(f"& \\scinum{{{mean_ll_comp}}} $\\pm$ \\scinum{{{np.std([res[res_type]['comp_ll'] for res in curr_res.values() if res_type in res])}}} ", end="")
print()
print("diff ", end="")
for data_name in data_names:
    curr_res = results[data_name][method]
    ll_diff = np.array([res[res_type]["ll"] for res in curr_res.values() if res_type in res]) - np.array([res[res_type]["comp_ll"] for res in curr_res.values() if res_type in res])
    print(f"& \\scinum{{{np.mean(ll_diff)}}} $\\pm$ \\scinum{{{np.std(ll_diff)}}} ", end="")
print()

method_order = [
    "DiCE + SPN",
    "VAE + SPN",
    "CH-CVAE ",
    "FACE_eps",
    "FACE_knn",
    "PROPLACE",
    'MIO + SPN',
    'LiCE_optimize',
    'LiCE_quartile',
    'LiCE_sample',
    'LiCE_median',
]
method_map = {
    "DiCE + SPN" : "DiCE",
    "VAE + SPN": "VAE",
    "CH-CVAE ": "C-CHVAE",
    "FACE_eps": "FACE ($\\epsilon$)",
    "FACE_knn": "FACE (knn)",
    "PROPLACE": "PROPLACE",
    'LiCE_optimize': "LiCE (optimize)",
    'LiCE_quartile': "LiCE (quartile)",
    'LiCE_sample': "LiCE (sample)",
    'LiCE_median': "LiCE (median)",
    'MIO + SPN': "MIO",
}
method_map_short = {k:v for k,v in method_map.items()}
method_map_short['LiCE_optimize'] = "LiCE (optim.)"
method_map_short['LiCE_quartile'] = "LiCE (quart.)"
method_map_short['LiCE_median'] = "LiCE (median)"
method_map_short['MIO + SPN'] = "MIO (\\uline{+spn})"
method_map_short["DiCE + SPN"] = "DiCE (\\uline{+spn})"
method_map_short['VAE + SPN'] = "VAE (\\uline{+spn})"

res_type = "valid"

for method in method_order:
    print("& " + method_map_short[method], end="")
    for data_name in ["gmsc", "adult", "credit"]:
        curr_res = results[data_name][method]
        spar = [res[res_type]["sparsity"] for res in curr_res.values() if res_type in res]
        nll = [-res[res_type]["ll"] for res in curr_res.values() if res_type in res]
        dist = [res[res_type]["distance"] for res in curr_res.values() if res_type in res]
        for d in [nll, dist, spar]:
            print(f" & \\scinumone{{{np.mean(d)}}} $\\pm$ \\scinumone{{{np.std(d)}}}", end="")
    print(" \\\\")

print("\n\nvalid actionable")
for method in method_order:
    print("&" + method_map_short[method], end="")
    for data_name in ["gmsc", "adult", "credit"]:
        curr_res = results[data_name][method]
        val = ["valid" in res for res in curr_res.values()]
        act = ["actionable" in res for res in curr_res.values()]
        for d in [val, act]:
            print(f" & \\scinumone{{{np.mean(d)*100}}}\\%", end="")
    print(" \\\\")


print("\n\nTIME")
for method in method_order:
    print(method_map[method], end="")
    for data_name in ["gmsc", "adult", "credit"]:
        curr_res = results[data_name][method]
        time = [res["time"] for res in curr_res.values()]
        for d in [time]:
            print(f" & \\scinum{{{np.median(d)}}}s", end="")
    print(" \\\\")


# ONLY SAME 
res_type = "valid"

method_order = [m for m in method_order if "sample" not in m and "quart" not in m]  

same_is = {}
for data_name in data_names:
    same_is[data_name] = []
    methods = method_order
    if data_name == "gmsc":
        methods = [m for m in methods if m != "VAE + SPN"]
    for i in results[data_name][methods[0]].keys():
        if all(res_type in results[data_name][method][i] for method in methods):
            same_is[data_name].append(i)
print({k: len(v) for k, v in same_is.items()})

for method in method_order:
    print(method_map_short[method], end="")
    for data_name in ["gmsc", "adult", "credit"]:
        if data_name == "gmsc" and method == "VAE + SPN":
            print(" & - & - & -", end="")
            continue
        curr_res = results[data_name][method]
        spar = [curr_res[i][res_type]["sparsity"] for i in same_is[data_name]]
        nll = [-curr_res[i][res_type]["ll"] for i in same_is[data_name]]
        dist = [curr_res[i][res_type]["distance"] for i in same_is[data_name]]
        for d in [nll, dist, spar]:
            print(f" & \\scinumone{{{np.mean(d)}}} $\\pm$ \\scinumone{{{np.std(d)}}}", end="")
    print(" \\\\")
